+++ monai/optimizers/lr_finder.py
@@ -9,11 +9,11 @@ import torch
 import torch.nn as nn
 from numpy.core.arrayprint import _none_or_positive_arg
 from torch.optim import Optimizer
-from torch.optim.lr_scheduler import _LRScheduler
 from torch.utils.data import DataLoader
 
 from monai.networks.utils import eval_mode
 from monai.utils import copy_to_device, optional_import
+from monai.optimizers.lr_scheduler import ExponentialLR, LinearLR
 
 plt, has_matplotlib = optional_import("matplotlib.pyplot")
 tqdm, has_tqdm = optional_import("tqdm")
@@ -523,42 +523,6 @@ class LearningRateFinder(object):
             plt.show()
 
         return ax
-
-
-class _LRSchedulerMONAI(_LRScheduler):
-    def __init__(self, optimizer: Optimizer, end_lr: float, num_iter: int, last_epoch: int = -1) -> None:
-        """
-        Args:
-            optimizer: wrapped optimizer.
-            end_lr: the final learning rate.
-            num_iter: the number of iterations over which the test occurs.
-            last_epoch: the index of last epoch.
-        Returns:
-            None
-        """
-        self.end_lr = end_lr
-        self.num_iter = num_iter
-        super(_LRSchedulerMONAI, self).__init__(optimizer, last_epoch)
-
-
-class LinearLR(_LRSchedulerMONAI):
-    """Linearly increases the learning rate between two boundaries over a number of
-    iterations.
-    """
-
-    def get_lr(self):
-        r = self.last_epoch / (self.num_iter - 1)
-        return [base_lr + r * (self.end_lr - base_lr) for base_lr in self.base_lrs]
-
-
-class ExponentialLR(_LRSchedulerMONAI):
-    """Exponentially increases the learning rate between two boundaries over a number of
-    iterations.
-    """
-
-    def get_lr(self):
-        r = self.last_epoch / (self.num_iter - 1)
-        return [base_lr * (self.end_lr / base_lr) ** r for base_lr in self.base_lrs]
 
 
 class StateCacher(object):
